started to work on adding command operators to script

+++ setup.sh
@@ -39,11 +39,14 @@ config-pin -q p9.21
 sudo apt-get install python-serial
 
 # Git config
-git config --global user.name "Brian McIlwain"
-git config --global user.email "bmcilw1@gmail.com"
-
-# Get .configs setup
-cd ~
-git clone https://github.com/bmcilw1/configs.git ~ && cp -r ~/configs/. ~/. && rm -rf ~/configs
-curl -fLo ~/.vim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-vim -c PlugInstall
+if [ $1="b" ]
+    then
+    git config --global user.name "Brian McIlwain"
+    git config --global user.email "bmcilw1@gmail.com"
+    
+    # Get .configs setup
+    cd ~
+    git clone https://github.com/bmcilw1/configs.git ~ && cp -r ~/configs/. ~/. && rm -rf ~/configs
+    curl -fLo ~/.vim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+    vim -c PlugInstall
+fi
